// src/enums/Enums.ts
export enum BrandOptions {
  Avis = "Avis",
  Filo = "AvisFilo",
  Budget = "Budget",
  Multinet = "Multinet",
  Default = "",
}

export enum BrandColors {
  Avis = "#ad0a27",
  AvisLight = "#ad0a27",
  AvisDark = "#d0052d",
  Filo = "#ad0a27",
  FiloDark = "#d0052d",
  Budget = "#00285f",
  BudgetDark = "#001a3d",
  Multinet = "#ffffff",
  MultinetButton = "#f4f5f7",
  //#eff7f3
}

export enum BrandIcons {
  Avis = "/media/avis-logo.svg",
  Filo = "/media/avis-filo-logo.svg",
  Budget = "/media/budget-logo.svg",
  Multinet = "/media/multinet-logo.svg",
}